flask-api/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
blinker==1.9.0
boto3==1.40.31
botocore==1.40.31
click==8.2.1
Deprecated==1.2.18
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==3.13
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rich==14.1.0
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
